Add parallel reduction to select the k neighbors
Sequential code that was used to find the k nearest neighbors from
the distance was abandoned in favor of a parallel one. The new code
performs parallel reduction to find the minimum value of the
distance matrix (the nearest neighbor), in three steps. Then this
value is excluded from the distance matrix, and, whith the same
procedure, we find the next nearest neighbor, until we find the
k nearest neighbors.